+++ example/android/app/src/main/kotlin/com/udara/developer/carouselsliderexample/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.udara.developer.carouselsliderexample
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity : FlutterActivity() {
-}