+++ example/lib/main.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_carousel_slider/carousel_slider.dart';
-import 'package:flutter_carousel_slider/carousel_slider_indicators.dart';
-import 'package:flutter_carousel_slider/carousel_slider_transforms.dart';
 
 void main() {
   runApp(MyApp());
@@ -46,7 +44,7 @@ class _MyHomePageState extends State<MyHomePage> {
     "G",
   ];
 
-  bool _isPlaying = false;
+  bool _isPlaying = true;
 
   CarouselSliderController _sliderController;