+++ lib/carousel_slider.dart
@@ -2,11 +2,14 @@ library fluttercarouselslider;
 
 import 'dart:async';
 
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 
 import 'carousel_slider_indicators.dart';
 import 'carousel_slider_transforms.dart';
+
+export './carousel_slider_indicators.dart';
+export './carousel_slider_transforms.dart';
 
 const _kMaxValue = 200000000000;
 const _kMiddleValue = 100000;
@@ -29,7 +32,11 @@ class CarouselSlider extends StatefulWidget {
     this.scrollDirection = Axis.horizontal,
     this.unlimitedMode = false,
     this.initialPage = 0,
+    this.onSlideChanged,
+    this.onSlideStart,
+    this.onSlideEnd,
     this.controller,
+    this.clipBehavior = Clip.hardEdge,
   })  : slideBuilder = null,
         itemCount = children.length,
         super(key: key);
@@ -50,7 +57,11 @@ class CarouselSlider extends StatefulWidget {
     this.scrollDirection = Axis.horizontal,
     this.unlimitedMode = false,
     this.initialPage = 0,
+    this.onSlideChanged,
+    this.onSlideStart,
+    this.onSlideEnd,
     this.controller,
+    this.clipBehavior = Clip.hardEdge,
   })  : children = null,
         super(key: key);
 
@@ -72,7 +83,10 @@ class CarouselSlider extends StatefulWidget {
   final ScrollPhysics scrollPhysics;
   final Axis scrollDirection;
   final int initialPage;
-
+  final ValueChanged<int>? onSlideChanged;
+  final VoidCallback? onSlideStart;
+  final VoidCallback? onSlideEnd;
+  final Clip clipBehavior;
   final CarouselSliderController? controller;
 
   @override
@@ -99,6 +113,12 @@ class CarouselSliderController {
       _state!._setAutoSliderEnabled(isEnabled);
     }
   }
+
+  dispose() {
+    if (_state != null && _state!.mounted) {
+      _state!.dispose();
+    }
+  }
 }
 
 class _CarouselSliderState extends State<CarouselSlider> {
@@ -113,19 +133,44 @@ class _CarouselSliderState extends State<CarouselSlider> {
     return Stack(
       children: <Widget>[
         if (widget.itemCount > 0)
-          PageView.builder(
-            itemCount: widget.unlimitedMode ? _kMaxValue : widget.itemCount,
-            controller: _pageController,
-            scrollDirection: widget.scrollDirection,
-            physics: widget.scrollPhysics,
-            itemBuilder: (context, index) {
-              final slideIndex = index % widget.itemCount;
-              Widget slide = widget.children == null ? widget.slideBuilder!(slideIndex) : widget.children![slideIndex];
-              return widget.slideTransform.transform(context, slide, index, _currentPage, _pageDelta, widget.itemCount);
+
+          ///Notification Listener added in order to capture Slide Start and Slide End events
+          NotificationListener<ScrollNotification>(
+            onNotification: (notification) {
+              if (notification is ScrollStartNotification) {
+                widget.onSlideStart?.call();
+              } else if (notification is ScrollEndNotification) {
+                widget.onSlideEnd?.call();
+              }
+              return true;
             },
+            child: PageView.builder(
+              onPageChanged: (val) {
+                widget.onSlideChanged?.call(val);
+              },
+              clipBehavior: widget.clipBehavior,
+              scrollBehavior: ScrollConfiguration.of(context).copyWith(
+                scrollbars: false,
+                overscroll: false,
+                dragDevices: {PointerDeviceKind.touch, PointerDeviceKind.mouse},
+              ),
+              itemCount: widget.unlimitedMode ? _kMaxValue : widget.itemCount,
+              controller: _pageController,
+              scrollDirection: widget.scrollDirection,
+              physics: widget.scrollPhysics,
+              itemBuilder: (context, index) {
+                final slideIndex = index % widget.itemCount;
+                Widget slide = widget.children == null
+                    ? widget.slideBuilder!(slideIndex)
+                    : widget.children![slideIndex];
+                return widget.slideTransform.transform(context, slide, index,
+                    _currentPage, _pageDelta, widget.itemCount);
+              },
+            ),
           ),
         if (widget.slideIndicator != null && widget.itemCount > 0)
-          widget.slideIndicator!.build(_currentPage! % widget.itemCount, _pageDelta, widget.itemCount),
+          widget.slideIndicator!.build(
+              _currentPage! % widget.itemCount, _pageDelta, widget.itemCount),
       ],
     );
   }
@@ -167,10 +212,12 @@ class _CarouselSliderState extends State<CarouselSlider> {
 
   void _initPageController() {
     _pageController?.dispose();
-    _pageController = new PageController(
+    _pageController = PageController(
       viewportFraction: widget.viewportFraction,
       keepPage: widget.keepPage,
-      initialPage: widget.unlimitedMode ? _kMiddleValue * widget.itemCount + _currentPage! : _currentPage!,
+      initialPage: widget.unlimitedMode
+          ? _kMiddleValue * widget.itemCount + _currentPage!
+          : _currentPage!,
     );
     _pageController!.addListener(() {
       setState(() {